+++ packages/mongodb-chatbot-server/src/routes/conversations/conversationsRouter.ts
@@ -3,12 +3,7 @@ import Router from "express-promise-router";
 import { rateLimit, Options as RateLimitOptions } from "express-rate-limit";
 import slowDown, { Options as SlowDownOptions } from "express-slow-down";
 import validateRequestSchema from "../../middleware/validateRequestSchema";
-import {
-  ChatLlm,
-  SystemPrompt,
-  ConversationCustomData,
-  ConversationsService,
-} from "mongodb-rag-core";
+import { ConversationCustomData, ConversationsService } from "mongodb-rag-core";
 import {
   CommentMessageRequest,
   makeCommentMessageRoute,
@@ -21,21 +16,17 @@ import {
 import {
   AddMessageRequest,
   AddMessageToConversationRouteParams,
+  GenerateResponse,
   makeAddMessageToConversationRoute,
 } from "./addMessageToConversation";
 import { requireRequestOrigin } from "../../middleware/requireRequestOrigin";
 import { NextFunction, ParamsDictionary } from "express-serve-static-core";
 import { requireValidIpAddress } from "../../middleware";
 import {
-  FilterPreviousMessages,
-  GenerateUserPromptFunc,
-} from "../../processors";
-import {
   GetConversationRequest,
   makeGetConversationRoute,
 } from "./getConversation";
 import { UpdateTraceFunc } from "./UpdateTraceFunc";
-import { GenerateResponse } from "../GenerateResponse";
 
 /**
   Configuration for rate limiting on the /conversations/* routes.
