+++ packages/mongodb-chatbot-server/src/routes/index.ts
@@ -1,3 +1,2 @@
 export * from "./conversations";
-export * from "./GenerateResponse";
 export * from "./legacyGenerateResponse";
